pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]

[tool.pyright]
typeCheckingMode = "strict"
reportUnknownMemberType = "none"

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]

[tool.ruff.lint]
extend-select = ["I"] 
fixable = ["ALL"] 